fix(autoware_overlay_rviz_plugin): fix conflicting header files
