+++ visualization/autoware_overlay_rviz_plugin/autoware_mission_details_overlay_rviz_plugin/include/autoware_mission_details_overlay_rviz_plugin/remaining_distance_time_display.hpp
@@ -14,7 +14,7 @@
 
 #ifndef REMAINING_DISTANCE_TIME_DISPLAY_HPP_
 #define REMAINING_DISTANCE_TIME_DISPLAY_HPP_
-#include "overlay_utils.hpp"
+#include "autoware_mission_details_overlay_rviz_plugin/overlay_utils.hpp"
 
 #include <QImage>
 #include <QString>
