+++ visualization/autoware_overlay_rviz_plugin/autoware_overlay_rviz_plugin/include/autoware_overlay_rviz_plugin/gear_display.hpp
@@ -14,7 +14,7 @@
 
 #ifndef GEAR_DISPLAY_HPP_
 #define GEAR_DISPLAY_HPP_
-#include "overlay_utils.hpp"
+#include "autoware_overlay_rviz_plugin/overlay_utils.hpp"
 
 #include <QImage>
 #include <QString>
