+++ visualization/autoware_overlay_rviz_plugin/autoware_overlay_rviz_plugin/include/autoware_overlay_rviz_plugin/turn_signals_display.hpp
@@ -14,7 +14,7 @@
 
 #ifndef TURN_SIGNALS_DISPLAY_HPP_
 #define TURN_SIGNALS_DISPLAY_HPP_
-#include "overlay_utils.hpp"
+#include "autoware_overlay_rviz_plugin/overlay_utils.hpp"
 
 #include <QImage>
 #include <QString>
