+++ visualization/autoware_overlay_rviz_plugin/autoware_overlay_rviz_plugin/src/steering_wheel_display.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "steering_wheel_display.hpp"
+#include "autoware_overlay_rviz_plugin/steering_wheel_display.hpp"
 
 #include <QFontDatabase>
 #include <QPainter>
