+++ visualization/autoware_overlay_rviz_plugin/autoware_overlay_rviz_plugin/src/traffic_display.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "traffic_display.hpp"
+#include "autoware_overlay_rviz_plugin/traffic_display.hpp"
 
 #include <QFontDatabase>
 #include <QPainter>
